Add Support For Graphmart Selection
This commit adds support for selecting the graphmart to run a query
against. This selection pane displays the names of the online graphmarts
and allows the user to set a focused graphmart by pressing enter in that
pane. It adds function to then run a query in the current buffer against
that graphmart or to simply set the focused graphmart.